Add more UI stuff, settings save / load and more

+++ stuntjumps/client/completions.lua
@@ -5,6 +5,10 @@ class "c_Completions" {
     end,
 
     load = function(self)
+        if not File.exists("completions.json") then
+            return
+        end
+
         local completionFile = File.open("completions.json", true)
         if completionFile then
             local data = completionFile:read(completionFile:getSize())
